Refactor authentication flow: restore getCurrentUser endpoint in UserController and update JwtUtils to include user details in token claims
